wifi: Fix no country code update if fail to set it before
When the country code was changed but failed to set it via client mode.
(For example: Wifi off but Scan Only Mode is on - client mode switch
to scan mode without supplicant), the Country Code should be treated
as not updated.

So new design will be treat CC changed only when CC is updated.
Other case, the CC should try to update when CMM is ready.

Bug: 347811880
Flag: EXEMPT bugfix
Test: atest -c FrameworksWifiTests
Test: Manual Test, see b/347811880#comment1 about reproduce steps.
Change-Id: Iadb969de43a59d46ff18f5a0983e8d29ebb0ac3a
Signed-off-by: tejas101k <tejassingh649@rediffmail.com>